• Extracted onMenuSelected method to parent class as this will be same across al classes providing sort+filter options
